Factor out reusable code
